+++ Client/_Imports.razor
@@ -9,4 +9,4 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using Order.Client.Components
-@using M = Order.Shared.Contract.Constants.Messages
+@using M = Order.Client.Pages.Messages;
